render/pyproject.toml
[tool.poetry]
name = "bricklens-render"
version = "0.1.0"
description = ""
authors = ["Matt Welsh <mdw@mdw.la>"]
packages = [
    { include = "bricklens/" },
]

[tool.poetry.dependencies]
python = ">=3.7,<3.9"
Pillow = "^7.2.0"
pyldraw = { git = "https://github.com/mdwelsh/python-ldraw", branch = "master" }
rich = "^10.4.0"
google-cloud-storage = "^1.42.1"

[tool.poetry.dev-dependencies]
pytest = "^3.4"
